# nummus/web/spec/api.yaml
openapi: 3.0.0
servers:
  - url: https://nummus/api
info:
  description: >-
    This is a server for nummus: A personal financial information aggregator and
    planning tool. Collects and categorizes transactions, manages budgets,
    tracks investments, calculates net worth, and predicts future performance.
  version: 0.1.0
  title: nummus API
  license:
    name: MIT
    url: https://github.com/WattsUp/nummus/blob/master/LICENSE
...
paths:
...
  /account:
    post:
      tags:
        - Account
      summary: Add a new Account to the portfolio
      description: ""
      operationId: nummus.web.controller_account.create
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Malformed request
      requestBody:
        $ref: "#/components/requestBodies/Account"
...
  /accounts:
    get:
      tags:
        - Account
      summary: Get Accounts of Portfolio
      description: ""
      operationId: nummus.web.controller_account.get_all
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/AccountCategory"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
                  count:
                    description: Number of matches to request query
                    type: number
                    format: integer
        "422":
          description: Invalid parameter supplied
  /asset:
    post:
      tags:
        - Asset
      summary: Add a new Asset to the portfolio
      description: ""
      operationId: nummus.web.controller_asset.create
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Malformed request
      requestBody:
        $ref: "#/components/requestBodies/Asset"
...
  /assets:
    get:
      tags:
        - Asset
      summary: Get Assets of Portfolio
      description: ""
      operationId: nummus.web.controller_asset.get_all
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/AssetCategory"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Asset"
                  count:
                    description: Number of matches to request query
                    type: number
                    format: integer
                  next_offset:
                    description: Offset to use to fetch next page, null
                      indicates no more pages available
                    type: integer
                    nullable: true
        "400":
          description: Malformed request
  /budget:
    post:
      tags:
        - Budget
      summary: Add a new Budget to the portfolio
      description: ""
      operationId: nummus.web.controller_budget.create
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          description: Malformed request
      requestBody:
        $ref: "#/components/requestBodies/Budget"
  /budget/{budgetUUID}:
    get:
      tags:
        - Budget
      summary: Find Budget by UUID
      description: Returns a single Budget
      operationId: nummus.web.controller_budget.get
      parameters:
        - $ref: "#/components/parameters/PathBudgetUUID"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          description: Malformed request
        "404":
          description: Budget not found
    put:
      tags:
        - Budget
      summary: Update an existing Budget
      description: ""
      operationId: nummus.web.controller_budget.update
      parameters:
        - $ref: "#/components/parameters/PathBudgetUUID"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          description: Malformed request
        "404":
          description: Budget not found
      requestBody:
        $ref: "#/components/requestBodies/Budget"
    delete:
      tags:
        - Budget
      summary: Delete an Budget
      description: Deletes the Budget
      operationId: nummus.web.controller_budget.delete
      parameters:
        - $ref: "#/components/parameters/PathBudgetUUID"
      responses:
        "204":
          description: Successful operation
        "400":
          description: Malformed request
        "404":
          description: Budget not found
  /budgets:
    get:
      tags:
        - Budget
      summary: Get Budgets of Portfolio
      description: ""
      operationId: nummus.web.controller_budget.get_all
      parameters:
        - $ref: "#/components/parameters/Start"
        - $ref: "#/components/parameters/End"
        - $ref: "#/components/parameters/SortByAge"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  budgets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Budget"
                  count:
                    description: Number of matches to request query
                    type: number
                    format: integer
                  next_offset:
                    description: Offset to use to fetch next page, null
                      indicates no more pages available
                    type: integer
                    nullable: true
        "400":
          description: Malformed request
  /transaction:
    post:
      tags:
        - Transaction
      summary: Add a new Transaction to the portfolio
      description: ""
      operationId: nummus.web.controller_transaction.create
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Malformed request
        "422":
          description: Missing at least one TransactionSplit
      requestBody:
        $ref: "#/components/requestBodies/Transaction"
  /transaction/{transactionUUID}:
    get:
      tags:
        - Transaction
      summary: Find Transaction by UUID
      description: Returns a single Transaction
      operationId: nummus.web.controller_transaction.get
      parameters:
        - $ref: "#/components/parameters/PathTransactionUUID"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Malformed request
        "404":
          description: Transaction not found
    put:
      tags:
        - Transaction
      summary: Update an existing Transaction
      description: ""
      operationId: nummus.web.controller_transaction.update
      parameters:
        - $ref: "#/components/parameters/PathTransactionUUID"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Malformed request
        "404":
          description: Transaction not found
        "422":
          description: Missing at least one TransactionSplit
      requestBody:
        $ref: "#/components/requestBodies/Transaction"
    delete:
      tags:
        - Transaction
      summary: Delete an Transaction
      description: Deletes the Transaction
      operationId: nummus.web.controller_transaction.delete
      parameters:
        - $ref: "#/components/parameters/PathTransactionUUID"
      responses:
        "204":
          description: Successful operation
        "400":
          description: Malformed request
        "404":
          description: Transaction not found
  /transactions:
    get:
      tags:
        - Transaction
      summary: Get Transactions of Portfolio
      description: Returns TransactionSplits matching criteria
      operationId: nummus.web.controller_transaction.get_all
      parameters:
        - $ref: "#/components/parameters/Start"
        - $ref: "#/components/parameters/End"
        - $ref: "#/components/parameters/SortByAge"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/TransactionCategory"
        - $ref: "#/components/parameters/TransactionSubcategory"
        - $ref: "#/components/parameters/TransactionTag"
        - $ref: "#/components/parameters/TransactionLocked"
        - $ref: "#/components/parameters/AccountUUID"
        - $ref: "#/components/parameters/AccountCategoryExplicit"
        - $ref: "#/components/parameters/AssetUUID"
        - $ref: "#/components/parameters/AssetCategoryExplicit"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/TransactionSplit"
                  next_offset:
                    description: Offset to use to fetch next page, null
                      indicates no more pages available
                    type: integer
                    nullable: true
        "400":
          description: Malformed request
        "404":
          description: Account or Asset not found
...
components:
  requestBodies:
    Account:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Account"
      description: Account object that needs to be added to the portfolio
      required: true
    Asset:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Asset"
      description: Asset object that needs to be added to the portfolio
      required: true
    Budget:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Budget"
      description: Budget object that needs to be added to the portfolio
      required: true
    Transaction:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Transaction"
      description: Transaction object that needs to be added to the portfolio
      required: true
...
  parameters:
...
    PathBudgetUUID:
      name: budgetUUID
      in: path
      description: ID of Budget to operate on
      required: true
      schema:
        type: string
        format: uuid
    PathTransactionUUID:
      name: transactionUUID
      in: path
      description: ID of Transaction to operate on
      required: true
      schema:
        type: string
        format: uuid
    Start:
      name: start
      in: query
      description: Filter data equal to or after this date
      schema:
        type: string
        format: date
    End:
      name: end
      in: query
      description: Filter data equal to or before this date, default is today
      schema:
        type: string
        format: date
    SortByAge:
      name: sort
      in: query
      description: Sort data by criteria
      schema:
        type: string
        default: oldest
        enum:
          - oldest
          - newest
...
    Limit:
      name: limit
      in: query
      description: Maximum number of elements to return
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    Offset:
      name: offset
      in: query
      description: Pagination offset to continue from
      schema:
        type: integer
        default: 0
    AccountCategory:
      name: category
      in: query
      description: Filter data by Account category
      schema:
        $ref: "#/components/schemas/Account/properties/category"
    AccountCategoryExplicit:
      name: account_category
      in: query
      description: Filter data by Account category
      schema:
        $ref: "#/components/schemas/Account/properties/category"
    AccountUUID:
      name: account
      in: query
      description: Filter data by Account UUID
      schema:
        $ref: "#/components/schemas/Account/properties/uuid"
    AssetCategory:
      name: category
      in: query
      description: Filter data by Asset category
      schema:
        $ref: "#/components/schemas/Asset/properties/category"
    AssetCategoryExplicit:
      name: asset_category
      in: query
      description: Filter data by Asset category
      schema:
        $ref: "#/components/schemas/Asset/properties/category"
    AssetUUID:
      name: asset
      in: query
      description: Filter data by Asset UUID
      schema:
        $ref: "#/components/schemas/Asset/properties/uuid"
    TransactionCategory:
      name: category
      in: query
      description: Filter data by Transaction category
      schema:
        $ref: "#/components/schemas/TransactionSplit/properties/category"
    TransactionSubcategory:
      name: subcategory
      in: query
      description: Filter data by Transaction subcategory
      schema:
        $ref: "#/components/schemas/TransactionSplit/properties/subcategory"
    TransactionTag:
      name: tag
      in: query
      description: Filter data by Transaction tag
      schema:
        $ref: "#/components/schemas/TransactionSplit/properties/tag"
    TransactionLocked:
      name: locked
      in: query
      description: Filter data by Transaction locked state
      schema:
        $ref: "#/components/schemas/Transaction/properties/locked"
    Search:
      name: search
      in: query
      description: >-
        After applying all other filters, perform a fuzzy search on data and
        discard poor matches. If sort is not provided, results are returned
        sorted by best match.
      schema:
        type: string
  schemas:
    Account:
      description: An financial Account of the portfolio
      type: object
      required:
        - uuid
        - name
        - institution
        - category
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: Monkey Bank Checking
        institution:
          type: string
          example: Monkey Bank
        category:
          type: string
          enum:
            - cash
            - credit
            - investment
            - mortgage
            - loan
            - fixed
            - other
        opened_on:
          description: Date of first transaction
          type: string
          format: date
          readOnly: true
        updated_on:
          description: Date of latest transaction
          type: string
          format: date
          readOnly: true
    Asset:
      description: Individual item with dynamic worth
      type: object
      required:
        - uuid
        - name
        - category
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        name:
          description: Name of asset
          type: string
          example: NASDAQ:AMZN
        description:
          type: string
          example: Amazon.com, Inc.
          nullable: true
        category:
          type: string
          enum:
            - cash
            - security
            - real_estate
            - vehicle
            - item
        unit:
          description: Unit name for a individual asset
          type: string
          example: shares
          nullable: true
        tag:
          type: string
          example: Collectibles
          nullable: true
    Budget:
      description: >-
        Allocation of cash flow transactions by categories. All numbers are
        annual allocations.
      type: object
      required:
        - uuid
        - date
        - categories
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        date:
          description: Effective date of budget
          type: string
          format: date
        total:
          description: Total limit of budget
          type: number
          format: double
          readOnly: true
        categories:
          description: Category breakdown of budget
          type: object
          properties:
            home:
              type: number
              format: double
            food:
              type: number
              format: double
            shopping:
              type: number
              format: double
            hobbies:
              type: number
              format: double
            services:
              type: number
              format: double
            travel:
              type: number
              format: double
          required:
            - home
            - food
            - shopping
            - hobbies
            - services
            - travel
    Transaction:
      description: Exchange of cash for an asset (or none)
      type: object
      required:
        - uuid
        - account_uuid
        - date
        - total
        - statement
        - locked
        - splits
        - is_split
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        account_uuid:
          description: ID of Account that owns this Transaction
          type: string
          format: uuid
        date:
          type: string
          format: date
        total:
          description: >-
            Total amount of cash exchanged. Positive indicates Account increases
            in value (inflow).
          type: number
          format: double
          example: -10.99
        statement:
          description: Text appearing on statement
          type: string
          example: AMZN#123456789
        locked:
          description: Parameters can only be manually edited
          type: boolean
        splits:
          description: Parent transaction when split
          type: array
          items:
            $ref: "#/components/schemas/TransactionSplit"
        is_split:
          description: True if there are more than one TransactionSplit
          type: boolean
          readOnly: true
    TransactionSplit:
      description: Exchange of cash for an asset (or none)
      type: object
      required:
        - uuid
        - total
        - parent_uuid
        - account_uuid
        - date
        - locked
        - is_split
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        parent_uuid:
          description: ID of Transaction that owns this TransactionSplit
          type: string
          format: uuid
          readOnly: true
        account_uuid:
          description: ID of Account that owns this Transaction
          type: string
          format: uuid
          readOnly: true
        date:
          type: string
          format: date
          readOnly: true
        total:
          description: >-
            Total amount of cash exchanged. Positive indicates Account increases
            in value (inflow).
          type: number
          format: double
          example: -10.99
        sales_tax:
          description: Amount of sales tax paid on transaction, always negative
          type: number
          format: double
          example: -0.52
          nullable: true
        payee:
          description: Name of payee (for outflow)/payer (for inflow)
          type: string
          example: Amazon
          nullable: true
        description:
          type: string
          example: Banana hook
          nullable: true
        category:
          type: string
          enum:
            - home
            - food
            - shopping
            - hobbies
            - services
            - travel
            - income
            - instrument
            - transfer
          nullable: true
        subcategory:
          type: string
          example: Furnishings
          nullable: true
        tag:
          type: string
          example: Kitchen
          nullable: true
        asset_uuid:
          description: >-
            ID of Asset exchanged for cash, primarily for instrument
            transactions
          type: string
          format: uuid
          nullable: true
        asset_quantity:
          description: >-
            Number of units of Asset exchanged. Positive indicates Account
            gained Assets (inflow)
          type: number
          format: double
          example: -0.52
          nullable: true
        locked:
          description: Parameters can only be manually edited
          type: boolean
          readOnly: true
        is_split:
          description: True if part of a split Transaction
          type: boolean
          readOnly: true
...